Refactors project and cleans up assets (#25)
* Updates AKI Inage in Readme

* Removes unused image and design files

Deletes several unused image, PSD, and Sketch files from the public directories to reduce the repository size and remove unnecessary assets.

Adds a rule to .gitignore to exclude the thumbnails directory.

* Removes unused hertwig-auer files

Deletes the hertwig-auer directory and its contents, including prompt files and related documentation. These files are no longer needed in the project.

* Bumps version to 2.2.0

Increments the package version to indicate new features or bug fixes.

* Updates project name and fixes start script

Renames the project to "aki" for consistency.
Updates the start script to reflect the new project name
and adjusts the delay for opening the browser.

Adds a BROKEN pdf examples folder with a README.

* Updates database with new pdfs and thumbnails

Updates the database with new PDF entries and corresponding thumbnails.
Also renames the example directory and removes old database entries.

This ensures the application uses the latest PDF documents and
their associated thumbnails, improving data accuracy and
the user experience.

---------

Co-authored-by: Chris D <listingslab@gmail.com>